myCart a product add korte parcil

+++ src/components/MyCartCard/MyCartCard.jsx
@@ -1,24 +1,40 @@
 /* eslint-disable react/prop-types */
-import { useEffect, useState } from "react";
 
-const MyCartCard = ({item}) => {
-    
+import Swal from "sweetalert2";
+
+const MyCartCard = ({item,setUser,user} ) => {
     const {name,img,brandName,price} =item
-  
-    
     const handleDelete=(name)=>{
-        fetch(`http://localhost:5000/cart/${name}`,{
-            method: 'DELETE'
-        })
-        .then(res=>res.json())
-        .then(data=>{
-            console.log(data);
-            if(data.deletedCount > 0){
-                alert('delet successfully')
+        
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/cart/${name}`,{
+                    method: 'DELETE'
+                })
+                .then(res=>res.json())
+                .then(data=>{
+                    console.log(data);  
+                    
+                    const remening= user.filter(remeningData=> remeningData.name !== name)
+                    setUser(remening)
+
+                    Swal.fire({
+                        icon: 'success',
+                        title: 'Product deleted successfully',
+                    
+                      })
+                })
             }
-          
-            
-        })
+          })
+
     }
     return (
         <div className="card card-side w-1/2 mx-auto mb-5 bg-base-100 shadow-xl">
@@ -28,7 +44,8 @@ const MyCartCard = ({item}) => {
             <p>brandName :   {brandName}</p>
             <p>{price}</p>
             <div className="card-actions justify-end">
-            <button onClick={()=>handleDelete(name)} className="btn bg-[red] text-white">Delete</button>
+            {/* <button onClick={()=>handleDelete(name)} className="btn bg-[red] text-white">Delete</button> */}
+            <button onClick={()=>handleDelete(name)} type="submit" className="btn bg-[red] text-white">Delete</button>
             </div>
         </div>
         </div>
